% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell2fire_run.R
\name{cell2fire_run}
\alias{cell2fire_run}
\title{Main Entry Point for Cell2FireR}
\usage{
cell2fire_run(input_args = commandArgs(trailingOnly = TRUE))
}
\arguments{
\item{input_args}{A character vector of arguments.}
}
\description{
Main Entry Point for Cell2FireR
}
\examples{
## get help.
cell2fire_run("-h")
## run demo data
cell2fire_run(c("--input-instance-folder", "../../data/Sub40x40/",
"--output-folder", "output/Sub40x40/",
"--ignitions", "1",
"--sim-years", "1"))
}
