% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGeneratorR.R
\name{DataGrids}
\alias{DataGrids}
\title{Title}
\usage{
DataGrids(in_folder, n_cells)
}
\arguments{
\item{in_folder}{folder path with ESRI ASCII raster files}

\item{n_cells}{number of cells}
}
\value{
a list of vectors with elevazion, slope etc...
}
\description{
Title
}
