% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGeneratorR.R
\name{Dictionary}
\alias{Dictionary}
\title{Dictionary
Reads fbp_lookup_table.csv and creates dictionaries for fuel types and colors}
\usage{
Dictionary(filename)
}
\arguments{
\item{filename}{file path with lookup table}
}
\value{
list with rows and colors
}
\description{
Dictionary
Reads fbp_lookup_table.csv and creates dictionaries for fuel types and colors
}
