% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGeneratorR.R
\name{ForestGrid}
\alias{ForestGrid}
\title{ForestGrid
Reads the ASCII file with the forest grid structure and returns an array with all the cells and grid dimensions nxm
Modified Feb 2018 by DLW to read the forest params (e.g. cell size) as well}
\usage{
ForestGrid(filename, dict_obj)
}
\arguments{
\item{filename}{file path with ASCII file with the forest grid structure}

\item{dict_obj}{fbp_lookup table converted to a list}
}
\value{
list with all the cells and grid dimensions nxm
}
\description{
ForestGrid
Reads the ASCII file with the forest grid structure and returns an array with all the cells and grid dimensions nxm
Modified Feb 2018 by DLW to read the forest params (e.g. cell size) as well
}
