% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGeneratorR.R
\name{GenDataFile}
\alias{GenDataFile}
\title{GenDataFile
Main wrapper function}
\usage{
GenDataFile(in_folder)
}
\arguments{
\item{in_folder}{folder where the Data.csv file should be generated}
}
\value{
data frame with generated data
}
\description{
GenDataFile
Main wrapper function
}
