% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGeneratorR.R
\name{GenerateDat}
\alias{GenerateDat}
\title{GenerateDat
Generates the final Data.csv}
\usage{
GenerateDat(g_fuel_type, elev, ps, saz, curing, in_folder)
}
\arguments{
\item{g_fuel_type}{numeric vector with fuel type code}

\item{elev}{numeric vector with elevation}

\item{ps}{numeric vector with slope}

\item{saz}{numeric vector with saz (azimuth in degrees)}

\item{curing}{numeric vector with curing}

\item{in_folder}{path to input folder}
}
\value{
dataframe with dictionaries for the fuel types and cells' colors
}
\description{
GenerateDat
Generates the final Data.csv
}
