% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fuel_key.R
\name{get_fuel_key}
\alias{get_fuel_key}
\title{get_fuel_key}
\usage{
get_fuel_key(fuel_string)
}
\arguments{
\item{fuel_string}{Character string representing the fuel model.}
}
\value{
A single character ("S", "K", "C", or "P"). Defaults to "S".
}
\description{
Map fuel strings to Cell2Fire single-character flags
}
