% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseInputs2remove.R
\name{InitCells}
\alias{InitCells}
\title{InitCells}
\usage{
InitCells(NCells, FTypes2, ColorsDict, CellsGrid4, CellsGrid3)
}
\arguments{
\item{NCells}{total number of cells to process}

\item{FTypes2}{List. Mapping of fuel type strings (e.g., "m1") to integer codes.}

\item{ColorsDict}{List. colors of .....}

\item{CellsGrid4}{...}

\item{CellsGrid3}{....}
}
\value{
a list with cell values
}
\description{
InitCells
}
