% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseInputs.R
\name{ParseInputs}
\alias{ParseInputs}
\title{ParseInputs}
\usage{
ParseInputs(args, parser = NULL)
}
\arguments{
\item{args}{character string of arguments to parse.}

\item{parser}{can pass a parser if set.}
}
\value{
vector of arguments
}
\description{
ParseInputs
}
