% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell2fire_run.R
\name{cell2fire_run}
\alias{cell2fire_run}
\title{Main Entry Point for Cell2FireR}
\usage{
cell2fire_run(input)
}
\arguments{
\item{input}{The input list/reactive values from Shiny}
}
\value{
The path to the results directory
}
\description{
Main Entry Point for Cell2FireR
}
\examples{
## get help.
}
