% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectCores.R
\name{detectCores}
\alias{detectCores}
\title{detectCores}
\usage{
detectCores(all.tests = FALSE, logical = TRUE)
}
\arguments{
\item{all.tests}{}

\item{logical}{}
}
\value{
number of cores
}
\description{
detectCores
}
\examples{
#
}
