% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.missing.R
\name{is.missing}
\alias{is.missing}
\title{is.missing}
\usage{
is.missing(x, checkPath = T)
}
\arguments{
\item{x}{variable or path}

\item{checkPath}{if to check for path}
}
\value{
TRUE or FALSE
}
\description{
Checks for validity of variable and path
}
