% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cell2fire.R
\name{run_cell2fire}
\alias{run_cell2fire}
\title{Run Cell2Fire Simulation}
\usage{
run_cell2fire(
  fuel,
  fuel_model,
  input_folder,
  out_folder,
  elevation = NULL,
  slope = NULL,
  saz = NULL,
  cur = NULL,
  crown = FALSE,
  cbh = NULL,
  cbd = NULL,
  ccf = NULL,
  hm = NULL,
  ignition_mode = NULL,
  ignition_file = NULL,
  ignition_point = NULL,
  ignition_radius = NULL,
  firebreaks = NULL,
  weather_mode = "0. Single weather file",
  wea_file = NULL,
  weather_weights = NULL,
  nsims = 1,
  rng_seed = 123,
  sim_threads = 1,
  fmc = 85,
  ldfmcs = "constant",
  outputs = NULL,
  c2f_bin_path = c2f_bin_pathEnv(),
  template_dir,
  dry = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{fuel}{Path to fuel raster.}

\item{fuel_model}{Character string specifying the fuel model (e.g., "0. Scott & Burgan").}

\item{input_folder}{Directory containing input files (like weathers).}

\item{out_folder}{Base directory for outputs.}

\item{elevation}{Path to elevation raster.}

\item{slope}{Path to slope raster in degreees.}

\item{saz}{Path to azimuth raster in degrees.}

\item{cur}{Path to curvature raster.}

\item{crown}{Logical; whether to use crown data.}

\item{cbh}{Path to Crown Base Height raster.}

\item{cbd}{Path to Crown Bulk Density raster.}

\item{ccf}{Path to Crown Canopy Fraction raster.}

\item{hm}{Path to Canopy Height Model raster.}

\item{ignition_mode}{Character string specifying ignition mode.}

\item{ignition_file}{Path to ignition probability map.}

\item{ignition_point}{Path to ignition points layer.}

\item{ignition_radius}{Numeric radius for ignition.}

\item{firebreaks}{Path to firebreaks raster.}

\item{weather_mode}{Character string specifying weather mode.}

\item{wea_file}{Path to single weather CSV.}

\item{weather_weights}{Path to weather weights file.}

\item{nsims}{Integer number of simulations.}

\item{rng_seed}{Integer seed for RNG.}

\item{sim_threads}{Integer number of threads.}

\item{fmc}{Numeric Fuel Moisture Content.}

\item{ldfmcs}{Character string for scenario.}

\item{outputs}{Character vector of output flags (e.g., c("grids", "stats")).}

\item{c2f_bin_path}{Path to the Cell2Fire executable. Defaults to the user data path, but user can specify a different path.}

\item{template_dir}{Path to the directory containing lookup tables and default weather.}

\item{dry}{Logical; if TRUE, only prepares files and returns the command arguments.}

\item{verbose}{Logical; if TRUE, provides many messages.}
}
\value{
A list object containing the process handle and metadata
for a fire spread simulation from cell2fire.
}
\description{
Run Cell2Fire Simulation
}
\section{Fields}{

\describe{
\item{\code{process}}{A \code{processx::process} object.}

\item{\code{command}}{Character string. Path to the binary.}

\item{\code{args}}{Character vector. The CLI arguments used.}
}}

